Consolidated real_fake_dataset.py into bitmind directory, updated references
